requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
tortoise-orm>=0.20.0
alembic>=1.12.0
aiosqlite>=0.19.0
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ruff>=0.1.0
httpx>=0.25.0

